fix(PromotionStrategy): fix promote cases and salary type

+++ src/main/java/net/ozgegurel/Employee.java
@@ -1,12 +1,13 @@
 package net.ozgegurel;
 
+import java.math.BigDecimal;
 
 public class Employee {
     public String name;
-    public Float salary;
+    public BigDecimal salary;
     public PromoteStrategy strategy;
     
-    public Employee(String name, Float salary, PromoteStrategy strategy){
+    public Employee(String name, BigDecimal salary, PromoteStrategy strategy){
         this.name = name;
         this.salary = salary;
         this.strategy = strategy;
